requirements.txt
appnope==0.1.2
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dask==2021.6.2
decorator==4.4.2
greenlet==1.1.0
html2text==2020.1.16
idna==2.10
ipython==7.20.0
ipython-genutils==0.2.0
jedi==0.18.0
lxml==4.6.2
multitasking==0.0.9
numpy==1.20.0
pandas==1.2.1
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
praw==7.2.0
prawcore==2.1.0
prawn==0.2.1
prompt-toolkit==3.0.14
psaw==0.1.0
ptyprocess==0.7.0
pyaml==20.4.0
Pygments==2.7.4
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.4.17
traitlets==5.0.5
update-checker==0.18.0
urllib3==1.26.3
wcwidth==0.2.5
websocket-client==1.1.0
yfinance==0.1.55